task/runner: Avoid re-running poison tasks
We have some tasks that are being re-executed
over and over again since they get stuck in the
task-runner logic. We should fix the root cause
of those, but to avoid the problem from getting
worse we should also avoid re-running these tasks
over and over again.

This fixes that by not even starting tasks that we
find out had already been started (phase=running).

This is related to #19