clients/ipfs: Use CIDv1 for all IPFS uploads
It is the recommended approach for CIDs as it's
future-proof (hash alg is embedded in the CID)
and also because it's case insensitive and will
work better on all scenarios (like URLs).

It is weirdly not the default on Piñata yet, but
while it is not we can easily toggle it on our
export requests.

It does not break compatibility on anything that I can
think of.

Reference: https://docs.ipfs.io/how-to/best-practices-for-ipfs-builders/