task: Update go-api-client and stream-tester with fixes (#43)
* go.mod: Update go-api-client to v0.2.0-beta

* go.mod: Update stream-tester to v0.12.11

Fixes error handlign on segmenter

* task: Update usage of go-api-client

* task: Fix cancellation handling on segmenter

If the loop finishes because ctx was cancelled, we'll
have no indication about the cancellation. Need to
check that explicitly.

Also notice that even if the segmenter sent the error on the
output channel, it would not be sufficient. That because if
the ctx is cancelled, it might end up never sending the error
anyway so we'd still need to check.

* task/transcode: Handle errors writing trailer

* go.mod: Use merged version of go-api-client (0.2.0)

* imports: Update all go-api-client imports to api