Handle the case where an asset is deleted by the user part way through processing
We were seeing lots of error building task context errors for one particular user, this was because they were deleting their asset immediately upon any task failures which meant any further callbacks would error on fetching the asset.

With this PR we'll now catch this specific case and return a different
error which we can add to the excusable list.